remove emmet-defparameter
It was first added as a helper macro in the test file, where it's useful
to work around eval-last-sexp etc. not changing the value of a variable
when eval'ing defvar. In commit
918e5420388bded0dd6b3656cc2fe9f4ba9c0649, it was moved to the main files
probably for the same purpose.

In Emacs 28, eval-last-sexp now redefines a defvar as expected, so this
wrapper is no longer useful.